Restructure distribution tests
